Former-commit-id: f9eef98690f606172c18c627b679c3186a8c9766

+++ include/igl/writeOFF.cpp
@@ -38,11 +38,11 @@ IGL_INLINE bool igl::writeOFF(
   
 //  for (int i = 0; i < F.rows(); i++)
 //      fprintf (fp, "3 %d %d %d\n", F(i,0), F(i,1), F(i,2));
-  for (int i = 0; i < F.rows(); i++)
+  for (int i = 0; i < (int)F.rows(); i++)
   {
-    fprintf (fp, "%d", F.cols());
-    for (int j = 0; j < F.cols(); j++)
-      fprintf (fp, " %d", F(i,j));
+    fprintf (fp, "%d", (int)F.cols());
+    for (int j = 0; j < (int)F.cols(); j++)
+      fprintf (fp, " %d", (int)F(i,j));
     fprintf (fp, "\n");
   }
   fclose (fp);
